+++ R/autospc.R
@@ -306,6 +306,11 @@ autospc <- function(data,
 
   arguments <- validate_argument_values(arguments)
 
+  arguments <- validate_arguments_against_data(
+    arguments,
+    x_values = data[[x_name]]
+  )
+
   arguments <- validate_algorithm_parameters(
     arguments,
     user_env = rlang::caller_env()